@@ -1,6 +1,6 @@
 ﻿namespace skullOS.Modules.Interfaces
 {
-    internal interface ISupport
+    internal interface ISupportModule
     {
     }
 }
